Removing "mkdir -p" and making sure that libraries are built in the build/root dir when making build decisions

+++ src/sam-template.js
@@ -3,14 +3,15 @@ const { exec, execSync } = require('child_process');
 const { EventEmitter } = require('events');
 const fs = require('fs');
 const yaml = require('js-yaml');
-const { folderUpdated, writeCacheFile, execOnlyShowErrors } = require('./tsc-tools');
+const { folderUpdated, writeCacheFile, execOnlyShowErrors, mkdir } = require('./tsc-tools');
 const path = require('path');
 const cfSchema = require('cloudformation-js-yaml-schema');
 const aws = require('aws-sdk');
 const archiver = require('archiver');
 
 const tempDir = "./.build/tmp";
-execOnlyShowErrors(`bash -c "mkdir -p ${tempDir}"`);
+mkdir(tempDir);
+
 let buildFlags = {};
 
 let stackeryConfig;
@@ -193,7 +194,8 @@ class SAMCompiledDirectory {
     }
 
     buildIfNotPresent() {
-        if(!fs.existsSync(this.outDir)) {
+        const outDir = `${buildRoot}/${this.tsconfigDir}/${this.outDir}`;
+        if(!fs.existsSync(outDir)) {
             this.build(undefined, true);
         }
     }
@@ -228,9 +230,14 @@ class SAMCompiledDirectory {
                     const localOutDir = path.resolve(this.tsconfigDir, this.outDir);
                     const outDir = path.resolve(process.cwd(), `${buildRoot}/${this.tsconfigDir}`);
                     if(fs.existsSync(localOutDir)) {
-                        execOnlyShowErrors(`bash -c "rm -R ${localOutDir}"`, { cwd: this.path })
+                        console.log('Removing out dir', localOutDir);
+                        fs.rmdirSync(localOutDir, { recursive: true, force: true });
                     }
-                    execOnlyShowErrors(`bash -c "mkdir -p ${outDir}"`, { cwd: this.path })
+                    console.log('Making final destination', outDir);
+                    if(!fs.existsSync(outDir)) {
+                        fs.mkdirSync(outDir, { recursive: true })
+                    }
+                    console.log('Compiling tsc', compileFlags);
                     execOnlyShowErrors(`npx tsc ${compileFlags}`, { cwd: this.path });
                     execOnlyShowErrors(`bash -c "cp -R ${this.outDir || '.'} ${outDir}"`, { cwd: this.path });
                 } else {
@@ -238,7 +245,6 @@ class SAMCompiledDirectory {
                     // if(fs.existsSync(outDir)) {
                     //     execOnlyShowErrors(`bash -c "rm -R ${outDir}"`, { cwd: this.path })
                     // }
-
                     const transpileOnly = samconfig.transpile_only == 'true'? '--transpile-only' : '';
 
                     execOnlyShowErrors(`npx tsc ${compileFlags} --outDir ${outDir}` + transpileOnly, { cwd: this.path });
@@ -319,9 +325,7 @@ class SAMLayer {
 
         console.log('samtsc: constructing build directory');
         const nodejsPath = `${buildRoot}/${this.path}/${this.packageFolder}`;
-        if(!fs.existsSync(nodejsPath)) {
-            execOnlyShowErrors(`bash -c "mkdir -p ${nodejsPath}"`);
-        }
+        mkdir(nodejsPath);
         fs.copyFileSync(packPath, nodejsPath + 'package.json');
 
         console.log('samtsc: installing dependencies');
@@ -498,7 +502,9 @@ class SAMTemplate {
         }
 
         const buildPath = `${buildRoot}/${this.path}`;
-        fs.unlinkSync(buildPath);
+        if(fs.existsSync(buildPath)) {
+            fs.unlinkSync(buildPath);
+        }
         fs.writeFileSync(buildPath, yaml.dump(template, { schema: cfSchema.CLOUDFORMATION_SCHEMA}));
         this.events.emit('template-update', this);
         writeCacheFile(this.path, true);
